[backend/fronted] Fix cyclic dependencies
