+++ opencti-platform/opencti-graphql/src/domain/stixObjectOrStixRelationship.ts
@@ -4,7 +4,7 @@ import { type EntityOptions, storeLoadById } from '../database/middleware-loader
 import { ABSTRACT_STIX_OBJECT, ABSTRACT_STIX_REF_RELATIONSHIP, ABSTRACT_STIX_RELATIONSHIP } from '../schema/general';
 import { FunctionalError, UnsupportedError } from '../config/errors';
 import { isStixRefRelationship, RELATION_CREATED_BY, RELATION_OBJECT_MARKING } from '../schema/stixRefRelationship';
-import { listThings, storeLoadByIdWithRefs, transformPatchToInput, updateAttributeFromLoadedWithRefs, validateCreatedBy, validateMarking } from '../database/middleware';
+import { listThings, storeLoadByIdWithRefs, transformPatchToInput, updateAttributeFromLoadedWithRefs, validateCreatedBy } from '../database/middleware';
 import { notify } from '../database/redis';
 import { BUS_TOPICS } from '../config/conf';
 import type { AuthContext, AuthUser } from '../types/user';
@@ -12,6 +12,7 @@ import { type StixRefRelationshipAddInput, type StixRefRelationshipsAddInput } f
 import type { BasicStoreCommon, BasicStoreObject } from '../types/store';
 import { schemaRelationsRefDefinition } from '../schema/schema-relationsRef';
 import { buildRelationData } from '../database/data-builder';
+import { validateMarking } from '../utils/access';
 
 type BusTopicsKeyType = keyof typeof BUS_TOPICS;
